Корректирует модули отрисовки модального окна формы и ее валидации
