+++ index.html
@@ -233,8 +233,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="node_modules/pristinejs/dist/pristine.js"></script>
-  <script src="node_modules/nouislider/distribute/nouislider.js"></script>
+  <script src="vendor/pristine/pristine.min.js"></script>
   <script src="js/main.js" type="module"></script>
 </body>
 </html>
