+++ js/upload-picture.js
@@ -1,5 +1,6 @@
-import {createFormListener, removeFormListener} from './upload-picture-form.js';
-import {createSlider, setupSlider} from './upload-picture-slider.js';
+import {validateUploadPictureForm} from './upload-picture-form-validation.js';
+
+const pictureUploadForm = document.querySelector('.img-upload__form');
 
 const pictureUploadInput = document.querySelector('#upload-file');
 const pictureUploadPreview = document.querySelector('.img-upload__preview img');
@@ -7,11 +8,6 @@ const pictureEdit = document.querySelector('.img-upload__overlay');
 const closeButton = document.querySelector('.img-upload__cancel');
 
 const pictureScaleInput = document.querySelector('.scale__control--value');
-const pictureScaleDownButton = document.querySelector('.scale__control--smaller');
-const pictureScaleUpButton = document.querySelector('.scale__control--bigger');
-
-const effectLevelSliderContainer = document.querySelector('.img-upload__effect-level');
-const effectLevelSlider = document.querySelector('.effect-level__slider');
 
 const effectsList = document.querySelector('.effects__list');
 const effectLevelInput = document.querySelector('.effect-level__value');
@@ -20,17 +16,9 @@ const noPictureFilterInput = document.querySelector('.effects__radio[value=none]
 const hashtagInput = document.querySelector('.img-upload__text .text__hashtags');
 const commentInput = document.querySelector('.img-upload__text .text__description');
 
-const onScaleDownButtonClick = () => {
-  if (parseInt(pictureScaleInput.value, 10) > 25) {
-    pictureScaleInput.value = `${parseInt(pictureScaleInput.value, 10) - 25}%`;
-    pictureUploadPreview.style.transform = `scale(${parseInt(pictureScaleInput.value, 10) / 100})`;
-  }
-};
-
-const onScaleUpButtonClick = () => {
-  if (parseInt(pictureScaleInput.value, 10) < 100) {
-    pictureScaleInput.value = `${parseInt(pictureScaleInput.value, 10) + 25}%`;
-    pictureUploadPreview.style.transform = `scale(${parseInt(pictureScaleInput.value, 10) / 100})`;
+const onUploadFormSubmit = (evt) => {
+  if (!validateUploadPictureForm()) {
+    evt.preventDefault();
   }
 };
 
@@ -38,8 +26,6 @@ const onEffectInputClick = (evt) => {
   if (evt.target.closest('.effects__radio')) {
     pictureUploadPreview.className = '';
     pictureUploadPreview.classList.add(`effects__preview--${evt.target.value}`);
-
-    setupSlider(evt.target.value);
   }
 };
 
@@ -54,9 +40,7 @@ const onDocumentKeydown = (evt) => {
 };
 
 const createListeners = () => {
-  pictureScaleDownButton.addEventListener('click', onScaleDownButtonClick);
-  pictureScaleUpButton.addEventListener('click', onScaleUpButtonClick);
-
+  pictureUploadForm.addEventListener('submit', onUploadFormSubmit);
   effectsList.addEventListener('change', onEffectInputClick);
 
   closeButton.addEventListener('click', onCloseButtonClick);
@@ -64,9 +48,6 @@ const createListeners = () => {
 };
 
 const removeListeners = () => {
-  pictureScaleDownButton.removeEventListener('click', onScaleDownButtonClick);
-  pictureScaleUpButton.removeEventListener('click', onScaleUpButtonClick);
-
   effectsList.removeEventListener('change', onEffectInputClick);
 
   closeButton.removeEventListener('click', onCloseButtonClick);
@@ -82,15 +63,11 @@ const defaultSetupPictureUpload = () => {
   pictureUploadPreview.className = 'effects__preview--none';
   hashtagInput.value = '';
   commentInput.value = '';
-
-  effectLevelSliderContainer.classList.add('hidden');
 };
 
 function openPictureUpload () {
   createListeners();
-  createFormListener();
   defaultSetupPictureUpload();
-  createSlider();
 
   document.body.classList.add('modal-open');
   pictureEdit.classList.remove('hidden');
@@ -98,9 +75,7 @@ function openPictureUpload () {
 
 function closePictureUpload () {
   removeListeners();
-  removeFormListener();
   defaultSetupPictureUpload();
-  effectLevelSlider.noUiSlider.destroy();
 
   pictureUploadInput.value = '';
   document.body.classList.remove('modal-open');
@@ -111,4 +86,4 @@ const initUploadPictureModule = () => {
   pictureUploadInput.addEventListener('change', openPictureUpload);
 };
 
-export {closePictureUpload, initUploadPictureModule};
+export {initUploadPictureModule};
