+++ package.json
@@ -19,9 +19,5 @@
   "scripts": {
     "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
     "lint": "eslint js/"
-  },
-  "dependencies": {
-    "nouislider": "^14.6.3",
-    "pristinejs": "^1.0.0"
   }
 }
